Somoaudience bidder update (#193)
* SomoAudience bidder update, initial commit

* Getting error - Incorrect bid request extension format for bidRequest with id tid

* Adjusted .json files for ApplicationTest

* Applied requested changes, improved application test

* Applied requested changes, improved application test, changed imp.bidFloor from Float to BigDecimal and fixed corresponding errors

* Checkstyle fix

+++ src/main/java/org/prebid/server/bidder/appnexus/AppnexusAdapter.java
@@ -191,10 +191,10 @@ public class AppnexusAdapter extends OpenrtbAdapter {
         return bannerBuilder.build();
     }
 
-    private static Float bidfloor(AppnexusParams params) {
+    private static BigDecimal bidfloor(AppnexusParams params) {
         final BigDecimal reserve = params.getReserve();
         return reserve != null && reserve.compareTo(BigDecimal.ZERO) > 0
-                ? reserve.floatValue()
+                ? reserve
                 : null;
     }
