+++ src/main/java/org/prebid/server/bidder/beachfront/model/BeachfrontSlot.java
@@ -3,6 +3,7 @@ package org.prebid.server.bidder.beachfront.model;
 import lombok.AllArgsConstructor;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @AllArgsConstructor(staticName = "of")
@@ -13,7 +14,7 @@ public class BeachfrontSlot {
 
     String id;
 
-    Float bidfloor;
+    BigDecimal bidfloor;
 
     List<BeachfrontSize> sizes;
 }
